refactor(loan): remove old loan-demo page, update routes + E2E tests
